CLEANUP: Remove redundant code and fix inference attention bug
- Deleted 12 redundant files (~3,500 lines):
  * Backup directories (app/ui/tabs/backup/, app/ui/tabs/old/)
  * Test files in root directory (7 files)
  * Redundant cllm_forward_complete.c implementation
  * Backup file cllm_pure_token.c.backup

- Fixed critical inference bug:
  * Attention was completely disabled in inference (commented out)
  * Now uses proper cllm_attention_forward() with buffer allocation
  * Generation quality should improve significantly
  * Matches training implementation

- Updated header files:
  * Removed declarations for deleted cllm_forward_complete.c functions
  * Clean API surface

- Verified proper code merging:
  * Training uses proper attention (line 618)
  * Backward pass includes attention gradients (lines 860-890)
  * Optimizer updates ALL parameters (lines 366-450)

- Build status: Clean compilation, 0 warnings, 0 errors
- All improvements properly merged into existing code
- No functionality regressions

Files deleted:
- app/ui/tabs/backup/ (3 files)
- app/ui/tabs/old/ (3 files)
- test_*.c (7 files)
- src/ai/cllm_forward_complete.c
- src/ai/cllm_pure_token.c.backup

Files modified:
- src/ai/cllm_inference.c (fixed attention)
- include/cllm_inference.h (removed dead declarations)
- app/cllm_integration.c (minor updates)
- app/ui/tabs/tab_training.c (minor updates)
- src/ai/cllm_training.c (documentation updates)

+++ app/ui/tabs/tab_training.c
@@ -525,7 +525,7 @@ void handle_training_tab_click(AppState* state, int x, int y) {
                 printf("  This model will be saved to models/saved_model.cllm after training\n");
             } else {
                 printf("=== USING EXISTING MODEL ===\n");
-                printf("  Vocab: %u tokens\n", state->cllm_model->vocab_size);
+                printf("  Vocab: %lu tokens\n", (unsigned long)state->cllm_model->vocab_size);
                 printf("  Layers: %u\n", state->cllm_model->num_layers);
                 printf("  Training will update this model's weights\n");
                 printf("  Model will be saved to models/saved_model.cllm after training\n");
